+++ src/env.rs
@@ -1,15 +1,16 @@
-use std::{collections::HashMap};
+use std::collections::HashMap;
+use std::rc::Rc;
 
 use crate::types::{Args, Ret, Type};
 
 #[derive(Debug)]
-pub struct Env<'a> {
+pub struct Env {
     data: HashMap<String, Type>,
-    outer: Option<&'a Env<'a>>,
+    outer: Option<Rc<Env>>,
 }
 
-impl<'a> Env<'a> {
-    pub fn new(outer: Option<&'a Env>, binds: &[&str], exprs: &[Type]) -> Env<'a> {
+impl Env {
+    pub fn new(outer: Option<Rc<Env>>, binds: &[&str], exprs: &[Type]) -> Env {
         if binds.len() != exprs.len() {
             panic!("`binds` and `exprs` must have the same length");
         }
@@ -24,7 +25,7 @@ impl<'a> Env<'a> {
         env
     }
 
-    pub fn new_default() -> Env<'a> {
+    pub fn new_default() -> Env {
         let mut env = Env::new(None, &[], &[]);
         env.set("+", Type::Fun(sum));
         env.set("-", Type::Fun(sub));
@@ -37,31 +38,26 @@ impl<'a> Env<'a> {
         self.data.insert(symbol.to_owned(), value);
     }
 
-    pub fn find(&self, symbol: &str) -> Option<&'a Env> {
+    fn find(&self, symbol: &str) -> Option<&Env> {
         match self.data.get(symbol) {
-            Some(_) => Some(&self),
-            None => {
-                match self.outer {
-                    Some(env) => env.find(symbol),
-                    None => None,
-                }
+            Some(_) => Some(self),
+            None => match self.outer {
+                Some(ref env) => env.find(symbol),
+                None => None,
             },
         }
     }
 
-    pub fn get(&self, symbol: &str) -> Result<Type, String>  {
+    pub fn get(&self, symbol: &str) -> Result<Type, String> {
         match self.find(symbol) {
-            Some(env) => {
-                match env.data.get(symbol) {
-                    Some(value) => Ok(value.clone()),
-                    None => Err(format!("Env should have the symbol '{}'", symbol)),
-                }
+            Some(env) => match env.data.get(symbol) {
+                Some(value) => Ok(value.clone()),
+                None => Err(format!("Env should have the symbol '{}'", symbol)),
             },
             None => Err(format!("Symbol '{}' not found in any environment", symbol)),
         }
     }
 }
-
 
 fn sum(args: Args) -> Ret {
     match (args.get(0), args.get(1)) {
@@ -69,7 +65,9 @@ fn sum(args: Args) -> Ret {
         (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a + *b as f64)),
         (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 + *b)),
         (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a + *b)),
-        _ => Err(String::from("'+' operation failed. Types must be numeric (Int or Float)")),
+        _ => Err(String::from(
+            "'+' operation failed. Types must be numeric (Int or Float)",
+        )),
     }
 }
 
@@ -79,7 +77,9 @@ fn sub(args: Args) -> Ret {
         (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a - *b as f64)),
         (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 - *b)),
         (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a - *b)),
-        _ => Err(String::from("'-' operation failed. Types must be numeric (Int or Float)")),
+        _ => Err(String::from(
+            "'-' operation failed. Types must be numeric (Int or Float)",
+        )),
     }
 }
 
@@ -89,7 +89,9 @@ fn mul(args: Args) -> Ret {
         (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a * *b as f64)),
         (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 * *b)),
         (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a * *b)),
-        _ => Err(String::from("'*' operation failed. Types must be numeric (Int or Float)")),
+        _ => Err(String::from(
+            "'*' operation failed. Types must be numeric (Int or Float)",
+        )),
     }
 }
 
@@ -99,6 +101,8 @@ fn div(args: Args) -> Ret {
         (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a / *b as f64)),
         (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 / *b)),
         (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a / *b)),
-        _ => Err(String::from("'/' operation failed. Types must be numeric (Int or Float)")),
+        _ => Err(String::from(
+            "'/' operation failed. Types must be numeric (Int or Float)",
+        )),
     }
 }