+++ tests/test_step2.rs
@@ -1,9 +1,11 @@
+use std::rc::Rc;
+
 use mal_rust;
 use mal_rust::env::Env;
 
 #[test]
 fn testing_evaluation_of_arithmetic_operations() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("(+ 1 2)", &mut env), "3");
     assert_eq!(mal_rust::rep("(+ 5 (* 2 3))", &mut env), "11");
     assert_eq!(mal_rust::rep("(- (+ 5 (* 2 3)) 3)", &mut env), "8");
@@ -21,7 +23,7 @@ fn testing_evaluation_of_arithmetic_operations() {
 
 #[test]
 fn test_invalid_function_name() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(
         mal_rust::rep("(abc 1 2 3)", &mut env),
         "Symbol 'abc' not found in any environment"
@@ -30,13 +32,13 @@ fn test_invalid_function_name() {
 
 #[test]
 fn testing_empty_list() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("()", &mut env), "()");
 }
 
 #[test]
 fn testing_repl_env() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("(+ 1 2)", &mut env), "3");
     assert_eq!(mal_rust::rep("(/ (- (+ 5 (* 2 3)) 3) 4)", &mut env), "2");
 }
@@ -44,7 +46,7 @@ fn testing_repl_env() {
 #[test]
 #[ignore]
 fn testing_evaluation_within_collection_literals() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("[1 2 (+ 1 2)]", &mut env), "[1 2 3]");
 
     assert_eq!(mal_rust::rep("{\"a\" (+ 7 8)}", &mut env), "{\"a\" 15}");
@@ -55,7 +57,7 @@ fn testing_evaluation_within_collection_literals() {
 #[test]
 #[ignore]
 fn check_that_evaluation_hasnt_broken_empty_collections() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("[]", &mut env), "[]");
 
     assert_eq!(mal_rust::rep("{}", &mut env), "{}");