+++ tests/test_step3.rs
@@ -1,9 +1,11 @@
+use std::rc::Rc;
+
 use mal_rust;
 use mal_rust::env::Env;
 
 #[test]
 fn testing_def() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("(def! x 3)", &mut env), "3");
     assert_eq!(mal_rust::rep("x", &mut env), "3");
     assert_eq!(mal_rust::rep("(def! x 4)", &mut env), "4");
@@ -14,7 +16,7 @@ fn testing_def() {
 
 #[test]
 fn verifying_symbols_are_case_sensitive() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("(def! mynum 111)", &mut env), "111");
     assert_eq!(mal_rust::rep("(def! MYNUM 222)", &mut env), "222");
     assert_eq!(mal_rust::rep("mynum", &mut env), "111");
@@ -23,7 +25,7 @@ fn verifying_symbols_are_case_sensitive() {
 
 #[test]
 fn check_that_error_aborts_def() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(
         mal_rust::rep("(abc 1 2 3)", &mut env),
         "Symbol 'abc' not found in any environment"
@@ -36,7 +38,7 @@ fn check_that_error_aborts_def() {
 
 #[test]
 fn testing_let() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("(let* (z 9) z)", &mut env), "9");
     assert_eq!(mal_rust::rep("(let* (x 9) x)", &mut env), "9");
     assert_eq!(mal_rust::rep("(let* (z (+ 2 3)) (+ 1 z))", &mut env), "6");
@@ -49,7 +51,7 @@ fn testing_let() {
 
 #[test]
 fn testing_outer_environment() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("(def! a 4)", &mut env), "4");
     assert_eq!(mal_rust::rep("(let* (q 9) q)", &mut env), "9");
     assert_eq!(mal_rust::rep("(let* (q 9) a)", &mut env), "4");
@@ -59,7 +61,7 @@ fn testing_outer_environment() {
 #[test]
 #[ignore]
 fn testing_let_with_vector_bindings() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("(let* [z 9] z)", &mut env), "9");
     assert_eq!(
         mal_rust::rep("(let* [p (+ 2 3) q (+ 2 p)] (+ p q))", &mut env),
@@ -70,7 +72,7 @@ fn testing_let_with_vector_bindings() {
 #[test]
 #[ignore]
 fn testing_vector_evaluation() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(
         mal_rust::rep("(let* (a 5 b 6) [3 4 a [b 7] 8])", &mut env),
         "[3 4 5 [6 7] 8]"
@@ -79,6 +81,6 @@ fn testing_vector_evaluation() {
 
 #[test]
 fn check_that_last_assignment_takes_priority() {
-    let mut env = Env::new_default();
+    let mut env = Rc::new(Env::new_default());
     assert_eq!(mal_rust::rep("(let* (x 2 x 3) x)", &mut env), "3");
 }