+++ changelog.rst
@@ -4,6 +4,7 @@ CHANGELOG
 0.6.0
 -----
 - Update `Env::new` to take two new parameters: `binds` and `exprs`
+- Split tests in multiple files
 
 0.5.0
 -----