+++ changelog.rst
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 - Update `Env::new` to take two new parameters: `binds` and `exprs`
 - Split tests in multiple files
+- Add support to printer to print function values
 
 0.5.0
 -----