+++ changelog.rst
@@ -1,6 +1,10 @@
 CHANGELOG
 =========
 
+0.6.1
+-----
+- Use Rc in Env outer environment instead of references with lifetime
+
 0.6.0
 -----
 - Update `Env::new` to take two new parameters: `binds` and `exprs`